+++ example/lib/src/api/js_api.dart
@@ -1,11 +1,10 @@
 import 'dart:async';
-import 'package:dsbridge_flutter/dsbridge_flutter.dart';
 import 'package:flutter_dsbridge_annotation/flutter_dsbridge_annotation.dart';
 
 part 'js_api.js.dart';
 
 @DSBridge(enableDebug: true)
-class JsApi {
+class JsApi extends _$JsApi {
   @dsBridge()
   String testSyn(dynamic msg) {
     print('msg=$msg');