+++ lib/flutter_dsbridge_annotation.dart
@@ -1,1 +1,2 @@
 export 'src/annotations/annotations.dart';
+export 'package:dsbridge_flutter/dsbridge_flutter.dart';